+++ sope-gdl1/GDLAccess/EOAttribute.m
@@ -84,7 +84,7 @@ static EONull   *null = nil;
 }
 
 // Is equal only if same name; used to make aliasing ordering stable
-- (unsigned)hash {
+- (NSUInteger)hash {
   return [self->name hash];
 }
